Create JuStars
In this game, your goal is to control the blue player using the spacebar and jump to catch the yellow stars. Each time you catch a star, your score increases. The game has 50 different levels, with the number of stars increasing as you progress. The player is subject to gravity, and you need to time your jumps to avoid falling off the ground.